move _get_blocked_template to common http utils file
