+++ ddtrace/appsec/utils.py
@@ -8,10 +8,9 @@ from ddtrace.appsec._constants import API_SECURITY
 from ddtrace.constants import APPSEC_ENV
 from ddtrace.internal.compat import parse
 from ddtrace.internal.compat import to_bytes_py2
-from ddtrace.internal.constants import APPSEC_BLOCKED_RESPONSE_HTML
-from ddtrace.internal.constants import APPSEC_BLOCKED_RESPONSE_JSON
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.utils.formats import asbool
+from ddtrace.internal.utils.http import _get_blocked_template  # noqa
 
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -80,54 +79,6 @@ def _appsec_rc_capabilities(test_tracer=None):
             result = str(base64.b64encode(value.to_bytes((value.bit_length() + 7) // 8, "big")), encoding="utf-8")
 
     return result
-
-
-_HTML_BLOCKED_TEMPLATE_CACHE = None  # type: Optional[str]
-_JSON_BLOCKED_TEMPLATE_CACHE = None  # type: Optional[str]
-
-
-def _get_blocked_template(accept_header_value):
-    # type: (str) -> str
-
-    global _HTML_BLOCKED_TEMPLATE_CACHE
-    global _JSON_BLOCKED_TEMPLATE_CACHE
-
-    need_html_template = False
-
-    if accept_header_value and "text/html" in accept_header_value.lower():
-        need_html_template = True
-
-    if need_html_template and _HTML_BLOCKED_TEMPLATE_CACHE:
-        return _HTML_BLOCKED_TEMPLATE_CACHE
-
-    if not need_html_template and _JSON_BLOCKED_TEMPLATE_CACHE:
-        return _JSON_BLOCKED_TEMPLATE_CACHE
-
-    if need_html_template:
-        template_path = os.getenv("DD_APPSEC_HTTP_BLOCKED_TEMPLATE_HTML")
-    else:
-        template_path = os.getenv("DD_APPSEC_HTTP_BLOCKED_TEMPLATE_JSON")
-
-    if template_path:
-        try:
-            with open(template_path, "r") as template_file:
-                content = template_file.read()
-
-            if need_html_template:
-                _HTML_BLOCKED_TEMPLATE_CACHE = content
-            else:
-                _JSON_BLOCKED_TEMPLATE_CACHE = content
-            return content
-        except (OSError, IOError) as e:
-            log.warning("Could not load custom template at %s: %s", template_path, str(e))  # noqa: G200
-
-    # No user-defined template at this point
-    if need_html_template:
-        _HTML_BLOCKED_TEMPLATE_CACHE = APPSEC_BLOCKED_RESPONSE_HTML
-        return APPSEC_BLOCKED_RESPONSE_HTML
-
-    _JSON_BLOCKED_TEMPLATE_CACHE = APPSEC_BLOCKED_RESPONSE_JSON
-    return APPSEC_BLOCKED_RESPONSE_JSON
 
 
 def parse_form_params(body):
